add split feature to card_area_organizer, fix bugs and add tests for card_area_organizer
